feat: update disease sections
