+++ packages/sections/src/disease/Ontology/index.ts
@@ -1,10 +1,10 @@
 import { lazy } from "react";
 
 export const definition = {
-  id: "knownDrugs",
-  name: "Known Drugs",
-  shortName: "KD",
-  hasData: data => data.knownDrugs?.count > 0 || data.knownDrugs.freeTextQuery || false,
+  id: "ontology",
+  name: "Ontology",
+  shortName: "O",
+  hasData: (data: any) => data.id,
 };
 
 // Components
