+++ packages/sections/src/disease/index.ts
@@ -1,0 +1,6 @@
+export * as Bibliography from "./Bibliography";
+export * as GWASStudies from "./GWASStudies";
+export * as KnownDrugs from "./KnownDrugs";
+export * as Ontology from "./Ontology";
+export * as OTProjects from "./OTProjects";
+export * as Phenotypes from "./Phenotypes";
