+++ packages/sections/src/index.tsx
@@ -1,5 +1,6 @@
 // Export the widgets
 export * as Drug from "./drug";
+export * as Disease from "./disease";
 
 // Export types
 export * from "./types/widget";
