+++ packages/ui/src/components/Summary/SummaryRenderer.tsx
@@ -1,5 +1,6 @@
 import { v1 } from "uuid";
 import { Widget } from "sections";
+import usePermissions from "../../hooks/usePermissions";
 
 type SummaryRendererProps = {
   widgets: Widget[];
@@ -7,20 +8,21 @@ type SummaryRendererProps = {
   keyPrefix?: string;
 };
 
-function SummaryRenderer({ 
-  widgets,
-  useKeys = true, 
-  keyPrefix = "summary" 
-}: SummaryRendererProps) {
+function SummaryRenderer({ widgets, useKeys = true, keyPrefix = "summary" }: SummaryRendererProps) {
+  const { isPartnerPreview } = usePermissions();
   return (
     <>
       {widgets.map((widget, index) => {
         const Summary = widget.Summary;
         const key = useKeys ? `${keyPrefix}-${v1()}` : `${keyPrefix}-${index}`;
+        // If the widget is private and we are not in partner preview, don't render it
+        if (widget.definition.isPrivate && !isPartnerPreview) {
+          return null;
+        }
         return <Summary key={key} />;
       })}
     </>
   );
 }
 
-export default SummaryRenderer; 
+export default SummaryRenderer;
