--- packages/ui/src/hooks/usePermissions.js
@@ -1,10 +1,0 @@
-import { getConfig } from "@ot/config";
-
-const config = getConfig();
-
-function usePermissions() {
-  const { isPartnerPreview } = config.profile;
-  return { isPartnerPreview };
-}
-
-export default usePermissions;
